requirements.txt
absl-py==1.4.0
addict==2.4.0
Augmentor==0.2.10
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==38.0.4
cycler==0.11.0
Cython==0.29.33
editdistance==0.6.2
Flask==2.2.2
flit_core==3.6.0
fonttools==4.38.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.1
idna==2.10
imageio==2.25.0
imgaug==0.4.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
natsort==8.2.0
networkx==2.6.3
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pip==22.3.1
pluggy==1.0.0
prettytable==3.6.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygame==2.1.2
pyOpenSSL==19.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scipy==1.7.3
setuptools==65.6.3
shapely==2.0.0
six==1.16.0
sortedcontainers==2.4.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminaltables==3.1.10
tifffile==2021.11.2
toolz==0.12.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.25.11
wcwidth==0.2.6
Werkzeug==2.2.2
wheel==0.35.1
yapf==0.32.0
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
zstandard==0.18.0
